+++ tests/cli/dorq/test_login.py
@@ -1,9 +1,13 @@
-from unittest.mock import MagicMock, Mock, PropertyMock, call
+from unittest.mock import MagicMock, Mock, call, create_autospec
 
 import pytest
 
 from orquestra.sdk._base.cli._dorq._login import _login
-from orquestra.sdk.schema.responses import ResponseStatusCode
+from orquestra.sdk._base.cli._dorq._repos import ConfigRepo, RuntimeRepo
+from orquestra.sdk._base.cli._dorq._ui._presenters import (
+    LoginPresenter,
+    WrappedCorqOutputPresenter,
+)
 
 
 class TestAction:
@@ -14,7 +18,8 @@ class TestAction:
     """
 
     @staticmethod
-    def test_passed_server_no_token():
+    @pytest.mark.parametrize("ce", [True, False])
+    def test_passed_server_no_token(ce):
         """
         Verifies how we pass variables between subcomponents.
         """
@@ -25,32 +30,32 @@ class TestAction:
         token = None
 
         # Mocks
-        exception_presenter = MagicMock()
-        login_presenter = MagicMock()
-        qe_repo = MagicMock()
-        config_repo = MagicMock()
-        qe_repo.get_login_url.return_value = config_url
+        exception_presenter = create_autospec(WrappedCorqOutputPresenter)
+        login_presenter = create_autospec(LoginPresenter)
+        runtime_repo = create_autospec(RuntimeRepo)
+        config_repo = create_autospec(ConfigRepo)
+
+        runtime_repo.get_login_url.return_value = config_url
 
         action = _login.Action(
             exception_presenter=exception_presenter,
             login_presenter=login_presenter,
-            qe_repo=qe_repo,
+            runtime_repo=runtime_repo,
             config_repo=config_repo,
         )
 
         # When
-        action.on_cmd_call(url=url, token=token)
+        action.on_cmd_call(url=url, token=token, ce=ce)
 
         # Then
         # We should get the login url from QE
-        assert qe_repo.mock_calls.count(call.get_login_url(url)) == 1
-        assert (
-            login_presenter.mock_calls.count(call.prompt_for_login(config_url, url))
-            == 1
-        )
+        assert runtime_repo.mock_calls.count(call.get_login_url(url, ce)) == 1
+        exception_presenter.show_error.assert_not_called()
+        login_presenter.prompt_for_login.assert_called_once_with(config_url, url, ce)
 
     @staticmethod
-    def test_passed_server_and_token():
+    @pytest.mark.parametrize("ce", [True, False])
+    def test_passed_server_and_token(ce):
         """
         Verifies how we pass variables between subcomponents.
         """
@@ -60,26 +65,25 @@ class TestAction:
         token = "my_token"
 
         config_name = "cfg"
-        exception_presenter = MagicMock()
-        login_presenter = MagicMock()
-        qe_repo = MagicMock()
-        config_repo = MagicMock()
+        exception_presenter = create_autospec(WrappedCorqOutputPresenter)
+        login_presenter = create_autospec(LoginPresenter)
+        runtime_repo = create_autospec(RuntimeRepo)
+        config_repo = create_autospec(ConfigRepo)
+
         config_repo.store_token_in_config.return_value = config_name
 
         action = _login.Action(
             exception_presenter=exception_presenter,
             login_presenter=login_presenter,
-            qe_repo=qe_repo,
+            runtime_repo=runtime_repo,
             config_repo=config_repo,
         )
 
         # When
-        action.on_cmd_call(url=url, token=token)
+        action.on_cmd_call(url=url, token=token, ce=ce)
 
         # Then
         # We should get the login url from QE
-        assert config_repo.mock_calls.count(call.store_token_in_config(url, token)) == 1
-        assert (
-            login_presenter.mock_calls.count(call.prompt_config_saved(url, config_name))
-            == 1
-        )
+        exception_presenter.show_error.assert_not_called()
+        config_repo.store_token_in_config.assert_called_once_with(url, token, ce)
+        login_presenter.prompt_config_saved.assert_called_once_with(url, config_name)