+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 👩‍🔬 *Experimental*
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry login`.
+* New CLI Login commands allows to login to CE with --ce flag
 
 🐛 *Bug Fixes*